+++ frontend/lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:place_prep/screens/auth_page.dart';
 import 'package:place_prep/screens/home_page.dart';
+import 'package:place_prep/services/auth_service.dart';
 
 void main() {
   runApp(const MyApp());
@@ -17,7 +19,7 @@ class MyApp extends StatelessWidget {
         useMaterial3: true,
         colorSchemeSeed: Colors.indigo,
       ),
-      home: const HomePage(),
+      home: AuthService.isAuthenticated ? const HomePage() : const AuthPage(),
     );
   }
 }